refactor: moved service_auto_update in the folder activities
